+++ scrollingpagerindicator/src/main/java/ru/tinkoff/scrollingpagerindicator/ScrollingPagerIndicator.java
@@ -5,6 +5,7 @@ import android.content.Context;
 import android.content.res.TypedArray;
 import android.graphics.Canvas;
 import android.graphics.Paint;
+import android.os.Build;
 import android.util.AttributeSet;
 import android.util.SparseArray;
 import android.view.View;
@@ -58,6 +59,7 @@ public class ScrollingPagerIndicator extends View {
 
     private Runnable attachRunnable;
     private PagerAttacher<?> currentAttacher;
+    private boolean autoRtl = true;
 
     private boolean dotCountInitialized;
 
@@ -314,6 +316,7 @@ public class ScrollingPagerIndicator extends View {
             currentAttacher.detachFromPager();
             currentAttacher = null;
             attachRunnable = null;
+            autoRtl = true;
         }
         dotCountInitialized = false;
     }
@@ -394,6 +397,17 @@ public class ScrollingPagerIndicator extends View {
         }
         adjustFramePosition(0, position);
         updateScaleInIdleState(position);
+    }
+
+    /**
+     * Sets Rtl direction availability when the view has Rtl direction.
+     * autoRtl is on by default.
+     *
+     * @param autoRtl false means rtl direction doesn't be apply even if view direction is Rtl.
+     */
+    public void setAutoRtl(boolean autoRtl) {
+        this.autoRtl = autoRtl;
+        invalidate();
     }
 
     @Override
@@ -538,7 +552,12 @@ public class ScrollingPagerIndicator extends View {
 
                 paint.setColor(calculateDotColor(scale));
                 if (orientation == LinearLayoutManager.HORIZONTAL) {
-                    canvas.drawCircle(dot - visibleFramePosition,
+                    float cx = dot - visibleFramePosition;
+                    if (autoRtl && isRtl()) {
+                        cx = getWidth() - cx;
+                    }
+
+                    canvas.drawCircle(cx,
                             getMeasuredHeight() / 2,
                             diameter / 2,
                             paint);
@@ -550,6 +569,11 @@ public class ScrollingPagerIndicator extends View {
                 }
             }
         }
+    }
+
+    private boolean isRtl() {
+        return Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR1 &&
+                getLayoutDirection() == LAYOUT_DIRECTION_RTL;
     }
 
     @ColorInt