+++ changelog.md
@@ -1,3 +1,12 @@
+## 1.2.2
+
+#### Fixes
+- support RTL direction in horizontal mode when the view is RTL [#29](https://github.com/Tinkoff/ScrollingPagerIndicator/issues/29)
+  - turn off RTL direction by `setAutoRtl(false)`
+
+#### Changes
+#### Additions
+
 ## 1.2.1
 
 #### Fixes